backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0
Flask-Mail>=0.9.1
pymongo>=4.6.0
cloudinary>=1.36.0
python-dotenv>=1.0.0
python-json-logger>=2.0.7
Werkzeug>=3.0.0
gunicorn>=21.2.0
requests>=2.31.0
pandas>=2.2.0
openpyxl>=3.1.2
numpy>=1.26.0
PyJWT>=2.8.0
bcrypt>=4.1.0
